Added a separate parent class to call on to check date validation
