--- lib/room.rb
@@ -1,15 +1,0 @@
-# require 'date'
-#
-# module Hotel
-#
-#   class Room
-#     attr_accessor :room_num, :booked_dates
-#
-#     def initialize(room_data)
-#       @room_num = room_data[:room_num]
-#       @booked_dates = room_data[:booked_dates]
-#     end
-#
-#   end
-#
-# end
